+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added Custom crafting table for armour
+### Added new texture for aquarium block
+
+## [1.19.3-1.0.0.alpha.4] - 2020-13-01
+
+### Fixed an issue where resources were not generating.
+
+## [1.19.3-1.0.0.alpha.3] - 2022-11-01
+
+### Bug Fixes
+
 ## [1.19.3-1.0.0.alpha.2] - 2022-07-01
 
 ### I fixed a lot of bugs.